Feature/permissions acp 0.10.0 20171125

+++ domain-services/jbb-permissions/src/main/java/org/jbb/permissions/impl/role/DefaultPermissionRoleService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 the original author or authors.
+ * Copyright (C) 2018 the original author or authors.
  *
  * This file is part of jBB Application Project.
  *
@@ -10,18 +10,16 @@
 
 package org.jbb.permissions.impl.role;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.Validate;
 import org.jbb.lib.eventbus.JbbEventBus;
 import org.jbb.permissions.api.PermissionRoleService;
-import org.jbb.permissions.api.entry.PermissionValue;
+import org.jbb.permissions.api.exceptions.RemovePredefinedRoleException;
 import org.jbb.permissions.api.matrix.PermissionTable;
 import org.jbb.permissions.api.permission.Permission;
 import org.jbb.permissions.api.permission.PermissionType;
+import org.jbb.permissions.api.permission.PermissionValue;
 import org.jbb.permissions.api.role.PermissionRoleDefinition;
+import org.jbb.permissions.api.role.PredefinedRole;
 import org.jbb.permissions.event.PermissionRoleChangedEvent;
 import org.jbb.permissions.event.PermissionRoleCreatedEvent;
 import org.jbb.permissions.event.PermissionRoleRemovedEvent;
@@ -38,12 +36,19 @@ import org.jbb.permissions.impl.role.model.AclRoleEntryEntity;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import lombok.RequiredArgsConstructor;
+
 @Service
 @RequiredArgsConstructor
 public class DefaultPermissionRoleService implements PermissionRoleService {
 
     private static final String ROLE_NOT_FOUND = "Role not found";
-    
+
     private final PermissionTypeTranslator permissionTypeTranslator;
     private final PermissionTranslator permissionTranslator;
     private final RoleTranslator roleTranslator;
@@ -61,14 +66,36 @@ public class DefaultPermissionRoleService implements PermissionRoleService {
     @Override
     public List<PermissionRoleDefinition> getRoleDefinitions(PermissionType permissionType) {
         AclPermissionTypeEntity permissionTypeEntity = permissionTypeTranslator
-            .toEntity(permissionType);
+                .toEntity(permissionType);
         return aclRoleRepository.findAllByPermissionTypeOrderByPositionAsc(permissionTypeEntity)
-            .stream().map(roleTranslator::toApiModel).collect(Collectors.toList());
+                .stream().map(roleTranslator::toApiModel).collect(Collectors.toList());
+    }
+
+    @Override
+    public PermissionRoleDefinition getRoleDefinition(Long roleId) {
+        Validate.notNull(roleId);
+        AclRoleEntity roleEntity = aclRoleRepository.findOne(roleId);
+        if (roleEntity != null) {
+            return roleTranslator.toApiModel(roleEntity);
+        } else {
+            return null;
+        }
+    }
+
+    @Override
+    public PermissionRoleDefinition getRoleDefinition(PredefinedRole predefinedRole) {
+        Validate.notNull(predefinedRole);
+        AclRoleEntity roleEntity = aclRoleRepository.findByPredefinedRole(predefinedRole);
+        if (roleEntity != null) {
+            return roleTranslator.toApiModel(roleEntity);
+        } else {
+            return null;
+        }
     }
 
     @Override
     public PermissionRoleDefinition addRole(PermissionRoleDefinition role,
-        PermissionTable permissionTable) {
+                                            PermissionTable permissionTable) {
         AclRoleEntity roleEntity = roleTranslator.toNewEntity(role);
         roleEntity = aclRoleRepository.save(roleEntity);
         putEntries(roleEntity, permissionTable);
@@ -79,57 +106,75 @@ public class DefaultPermissionRoleService implements PermissionRoleService {
 
     private void putEntries(AclRoleEntity roleEntity, PermissionTable permissionTable) {
         permissionTable.getPermissions().stream()
-            .map(permission -> roleEntryTranslator.toEntity(permission, roleEntity))
-            .forEach(aclRoleEntryRepository::save);
+                .map(permission -> roleEntryTranslator.toEntity(permission, roleEntity))
+                .forEach(aclRoleEntryRepository::save);
     }
 
     @Override
     public void removeRole(Long roleId) {
-        permissionCaches.clearCaches();
-        aclRoleRepository.delete(roleId);
-        eventBus.post(new PermissionRoleRemovedEvent(roleId));
+        Validate.notNull(roleId);
+        AclRoleEntity role = aclRoleRepository.findOne(roleId);
+        if (role != null) {
+            if (role.getPredefinedRole() != null) {
+                throw new RemovePredefinedRoleException();
+            }
+            permissionCaches.clearCaches();
+            aclRoleRepository.delete(roleId);
+            eventBus.post(new PermissionRoleRemovedEvent(roleId));
+        }
     }
 
     @Override
     public PermissionTable getPermissionTable(Long roleId) {
         AclRoleEntity roleEntity = Optional.ofNullable(aclRoleRepository.findOne(roleId))
-            .orElseThrow(() -> new IllegalArgumentException(ROLE_NOT_FOUND));
+                .orElseThrow(() -> new IllegalArgumentException(ROLE_NOT_FOUND));
+        return getPermissionTable(roleEntity);
+    }
+
+    @Override
+    public PermissionTable getPermissionTable(PredefinedRole predefinedRole) {
+        Validate.notNull(predefinedRole);
+        AclRoleEntity roleEntity = aclRoleRepository.findByPredefinedRole(predefinedRole);
+        return getPermissionTable(roleEntity);
+    }
+
+    private PermissionTable getPermissionTable(AclRoleEntity roleEntity) {
         List<AclRoleEntryEntity> roleEntries = aclRoleEntryRepository
-            .findAllByRole(roleEntity, new Sort("permission.position"));
+                .findAllByRole(roleEntity, new Sort("permission.position"));
         return permissionTableTranslator.fromRoleToApiModel(roleEntries);
     }
 
     @Override
     public PermissionRoleDefinition updateRoleDefinition(PermissionRoleDefinition role) {
         AclRoleEntity roleEntity = roleTranslator.toEntity(role)
-            .orElseThrow(() -> new IllegalArgumentException(ROLE_NOT_FOUND));
+                .orElseThrow(() -> new IllegalArgumentException(ROLE_NOT_FOUND));
         PermissionRoleDefinition roleDefinition = roleTranslator.toApiModel(
-            aclRoleRepository.save(roleEntity));
+                aclRoleRepository.save(roleEntity));
         eventBus.post(new PermissionRoleChangedEvent(roleDefinition.getId()));
         return roleDefinition;
     }
 
     @Override
     public PermissionTable updatePermissionTable(Long roleId,
-        PermissionTable permissionTable) {
+                                                 PermissionTable permissionTable) {
         permissionCaches.clearCaches();
         AclRoleEntity roleEntity = Optional.ofNullable(aclRoleRepository.findOne(roleId))
-            .orElseThrow(() -> new IllegalArgumentException(ROLE_NOT_FOUND));
+                .orElseThrow(() -> new IllegalArgumentException(ROLE_NOT_FOUND));
 
         Set<Permission> permissions = permissionTable.getPermissions();
         Set<AclPermissionEntity> permissionEntities = permissions.stream()
-            .map(permissionTranslator::toEntity)
-            .collect(Collectors.toSet());
+                .map(permissionTranslator::toEntity)
+                .collect(Collectors.toSet());
 
         Set<AclRoleEntryEntity> roleEntries = permissionEntities.stream()
-            .map(permissionEntity -> aclRoleEntryRepository
-                .findByRoleAndPermission(roleEntity, permissionEntity)
-                .orElse(AclRoleEntryEntity.builder()
-                    .role(roleEntity)
-                    .permission(permissionEntity)
-                    .build()
-                )
-            ).collect(Collectors.toSet());
+                .map(permissionEntity -> aclRoleEntryRepository
+                        .findByRoleAndPermission(roleEntity, permissionEntity)
+                        .orElse(AclRoleEntryEntity.builder()
+                                .role(roleEntity)
+                                .permission(permissionEntity)
+                                .build()
+                        )
+                ).collect(Collectors.toSet());
 
         roleEntries.forEach(entry -> updatePermissionValue(entry, permissions));
         aclRoleEntryRepository.save(roleEntries);
@@ -137,10 +182,10 @@ public class DefaultPermissionRoleService implements PermissionRoleService {
     }
 
     private void updatePermissionValue(AclRoleEntryEntity entry,
-        Set<Permission> permissions) {
+                                       Set<Permission> permissions) {
         PermissionValue newPermissionValue = permissions.stream()
-            .filter(p -> p.getDefinition().getCode().equals(entry.getPermission().getCode()))
-            .map(Permission::getValue).findFirst().orElse(PermissionValue.NO);
+                .filter(p -> p.getDefinition().getCode().equals(entry.getPermission().getCode()))
+                .map(Permission::getValue).findFirst().orElse(PermissionValue.NO);
         entry.setEntryValue(newPermissionValue);
     }
